Removed test fixture duplication
